Move getThemeFilterTerm
