+++ client/state/selectors/get-theme-showcase-description.js
@@ -6,8 +6,7 @@ import { get, includes } from 'lodash';
 /**
  * Internal dependencies
  */
-import { findThemeFilterTerm } from 'state/themes/selectors';
-import getThemeFilterTerm from 'state/selectors/get-theme-filter-term';
+import { findThemeFilterTerm, getThemeFilterTerm } from 'state/themes/selectors';
 
 import 'state/themes/init';
