+++ client/state/themes/selectors/get-theme-filter-term.js
@@ -18,6 +18,6 @@ import 'state/themes/init';
  * @param  {string}  term   The term slug
  * @returns {object}         A filter term object
  */
-export default function getThemeFilterTerm( state, filter, term ) {
+export function getThemeFilterTerm( state, filter, term ) {
 	return get( getThemeFilterTerms( state, filter ), term );
 }
