+++ client/state/themes/selectors/index.js
@@ -12,6 +12,7 @@ export { getThemeCustomizeUrl } from 'state/themes/selectors/get-theme-customize
 export { getThemeDemoUrl } from 'state/themes/selectors/get-theme-demo-url';
 export { getThemeDetailsUrl } from 'state/themes/selectors/get-theme-details-url';
 export { getThemeFilterStringFromTerm } from 'state/themes/selectors/get-theme-filter-string-from-term';
+export { getThemeFilterTerm } from 'state/themes/selectors/get-theme-filter-term';
 export { getThemeFilterTermFromString } from 'state/themes/selectors/get-theme-filter-term-from-string';
 export { getThemeForumUrl } from 'state/themes/selectors/get-theme-forum-url';
 export { getThemeHelpUrl } from 'state/themes/selectors/get-theme-help-url';
